+++ .eslintrc
@@ -14,7 +14,6 @@
     "browser": true
   },
   "globals": {
-    "adsbygoogle": true,
-    "graphql": true
+    "adsbygoogle": true
   }
 }
