+++ .gitignore
@@ -18,7 +18,6 @@ Thumbs.db
 # Project Specific
 node_modules
 npm-debug.log
-.cache/
 public/
 .env*
 .eslintcache
