+++ netlify.toml
@@ -22,5 +22,5 @@
   GATSBY_BUNGIE_API_KEY = "e44a1c455b634479aa069be92838857d"
   GATSBY_SITE_URL = "https://destinyclanwarfare.com"
   GATSBY_DISALLOW_ROBOTS = "false"
-  GATSBY_ENABLE_MATCH_HISTORY = "false"
+  GATSBY_ENABLE_MATCH_HISTORY = "true"
   GATSBY_ENABLE_PREVIOUS_LEADERBOARDS = "true"
