+++ package.json
@@ -9,13 +9,13 @@
     "yarn": "^1.2.1"
   },
   "scripts": {
-    "start": "npm-run-all --parallel start:gatsby",
-    "start:gatsby": "node --max-old-space-size=8192 ./node_modules/.bin/gatsby develop -p 9000",
-    "start:storybook": "start-storybook -p 9001 -c .storybook",
-    "build": "npm-run-all build:gatsby",
-    "build:gatsby": "node --max-old-space-size=8192 ./node_modules/.bin/gatsby build",
-    "build:storybook": "build-storybook -c .storybook -o ./public/storybook",
-    "clean": "rm -rf .cache && rm -rf public",
+    "start": "npm-run-all start:dev",
+    "start:dev": "react-static start",
+    "build": "npm-run-all build:production",
+    "build:staging": "npm run build:production -- --staging",
+    "build:production": "node --max-old-space-size=8192 ./node_modules/.bin/react-static build",
+    "clean": "rm -rf public",
+    "serve": "serve ./public -p 9000",
     "optimize": "npm-run-all --parallel optimize:*",
     "optimize:svg": "find ./src -type f -name '*.svg' -print0 | xargs -0 -n 1 -P 6 svgo --enable=removeXMLNS --enable=removeTitle --enable=removeStyleElement --multipass",
     "optimize:images": "imagemin src/images/* --out-dir=src/images",
@@ -23,31 +23,20 @@
     "qa": "npm-run-all --parallel qa:*",
     "qa:lint": "eslint --cache -f unix .",
     "qa:stylint": "stylint ./src",
-    "test": "npm-run-all --parallel qa",
+    "test": "npm-run-all qa",
     "deploy": "npm-run-all build deploy:firebase",
     "deploy:firebase": "firebase deploy --non-interactive --token $FIREBASE_TOKEN"
   },
   "dependencies": {
-    "@jacobmischka/gatsby-plugin-react-svg": "^1.0.0",
     "autoprefixer": "^7.2.3",
+    "autoprefixer-stylus": "^0.14.0",
     "axios": "^0.17.0",
+    "babel-polyfill": "^6.26.0",
     "camelcase-keys": "^4.1.0",
     "classnames": "^2.2.5",
-    "gatsby": "^1.9.127",
-    "gatsby-link": "^1.6.22",
-    "gatsby-plugin-canonical-urls": "^1.0.8",
-    "gatsby-plugin-create-client-paths": "^1.0.7",
-    "gatsby-plugin-feed": "^1.3.15",
-    "gatsby-plugin-google-analytics": "^1.0.11",
-    "gatsby-plugin-manifest": "^1.0.8",
-    "gatsby-plugin-netlify": "^1.0.18",
-    "gatsby-plugin-no-sourcemaps": "^1.0.3",
-    "gatsby-plugin-nprogress": "^1.0.10",
-    "gatsby-plugin-react-helmet": "^1.0.8",
-    "gatsby-plugin-react-next": "^1.0.7",
-    "gatsby-plugin-remove-serviceworker": "^1.0.0",
-    "gatsby-plugin-sitemap": "^1.2.5",
-    "gatsby-plugin-stylus": "^1.1.10",
+    "css-loader": "^0.28.11",
+    "dotenv": "^5.0.1",
+    "extract-text-webpack-plugin": "^3.0.2",
     "linkify-urls": "^1.4.0",
     "moment": "^2.19.2",
     "multi-sort": "^1.0.1",
@@ -56,12 +45,19 @@
     "pascal-case": "^2.0.1",
     "poststylus": "^1.0.0",
     "query-string": "^5.0.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-router": "^4.2.0",
+    "react-static": "^5.6.1",
     "responsive-grid": "^1.1.0",
     "sentence-case": "^2.1.1",
+    "style-loader": "^0.20.3",
+    "stylus": "^0.54.5",
+    "stylus-loader": "^3.0.2",
     "stylus-mixins": "^0.4.0",
+    "svg-react-loader": "^0.4.5",
     "svgo": "^1.0.3",
-    "uuid": "^3.1.0",
-    "yarn": "^1.3.2"
+    "uuid": "^3.1.0"
   },
   "devDependencies": {
     "eslint": "^3.19.0",
@@ -76,6 +72,7 @@
     "firebase-tools": "^3.17.4",
     "imagemin-cli": "^3.0.0",
     "npm-run-all": "^4.1.1",
+    "serve": "^6.1.0",
     "stylint": "^1.5.9"
   }
 }
