+++ src/components/advert/Advert.styl
@@ -15,4 +15,5 @@
 
   &#___gatsby
   &#nprogress
+  &#root
     display block !important // @stylint ignore
