+++ src/components/button/Button.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Link from 'gatsby-link'
+import { Link } from 'react-static'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
