+++ src/pages/404.js
@@ -1,11 +1,27 @@
 import React, { Component } from 'react'
-import Link from 'gatsby-link'
+import { Link } from 'react-static'
 import Helmet from 'react-helmet'
 import HoldingPage from '../components/holding-page/HoldingPage'
 import { Logo } from '../components/logo/Logo'
 
 class NotFoundPage extends Component {
+  constructor (props) {
+    super()
+
+    this.state = { active: false }
+  }
+
+  componentDidMount () {
+    this.setState({
+      active: true
+    })
+  }
+
   render () {
+    const { active } = this.state
+
+    if (!active) return null
+
     const title = 'Page not found'
     const description = 'Sorry, this page could not be found'
