+++ src/pages/clans.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import { withRouteData } from 'react-static'
 import PropTypes from 'prop-types'
 import Helmet from 'react-helmet'
 import PageContainer from '../components/page-container/PageContainer'
@@ -11,14 +12,16 @@ const constants = require('../utils/constants')
 class ClansPage extends Component {
   render () {
     const { data } = this.props
-    const leaderboard = data.allClan.edges.map(({ node }) => {
+    const leaderboard = data.allClan.map(clan => {
       return {
-        ...node,
-        clanId: `${constants.prefix.hash}${node.clanId}`
+        ...clan,
+        clanTag: clan.tag,
+        clanId: `${constants.prefix.hash}${clan.id}`
       }
     })
     const title = 'Clans'
     const description = 'All clans battling their way to the top of the Destiny 2 clan leaderboard'
+    const leaderboardColumns = [ 'color', 'foreground', 'background', 'platforms', 'name', 'clanTag', 'clanId' ]
 
     return (
       <PageContainer>
@@ -31,7 +34,7 @@ class ClansPage extends Component {
         <Card cutout center>
           <Lockup primary center kicker="All" heading="Clans" />
         </Card>
-        <Leaderboard cutout data={leaderboard} />
+        <Leaderboard cutout data={leaderboard} columns={leaderboardColumns} />
       </PageContainer>
     )
   }
@@ -41,33 +44,4 @@ ClansPage.propTypes = {
   data: PropTypes.object
 }
 
-export default ClansPage
-
-export const pageQuery = graphql`
-  query ClansPageQuery {
-    allClan(sort: { fields: [ nameSortable ] }) {
-      edges {
-        node {
-          path
-          platforms {
-            id
-            size
-            active
-          }
-          name
-          color
-          clanTag: tag
-          clanId: id
-          foreground {
-            color
-            icon
-          }
-          background {
-            color
-            icon
-          }
-        }
-      }
-    }
-  }
-`
+export default withRouteData(ClansPage)
